Refactor IntegerTime and IntegerTimeInterval into separate files
