+++ frontend/src/models/RehearsalSong.js
@@ -1,5 +1,6 @@
 import {Record} from 'immutable';
-import {IntegerTimeInterval, IntegerTime} from '@Utils/DateTimeUtils.js';
+import IntegerTime from '@Utils/IntegerTime.js';
+import IntegerTimeInterval from '@Utils/IntegerTimeInterval.js';
 import {isNumber} from '@Utils/TypeChecks.js';
 export default class RehearsalSong extends Record({id:-1, title:'', rehearsalTime: IntegerTimeInterval.fromArmyTime(0,1)}){
     setRehearsalTime(start, end){
